jmh/run.sh: make mvn clean optional
This commit makes cleaning of JARs used for the benchmarks optional.
In most of cases Maven should do a pretty good job of building a new JAR
without the need for cleaning the target directory.

This cuts the time it takes to start benchmarks from ~40s to ~30s
on my computer

+++ jmh/run.sh
@@ -1,24 +1,36 @@
 #!/bin/bash
 BASEDIR=$(dirname $0)
 
+CLEAN=""
+if [[ "$1" = "-clean" ]]; then
+    echo "Preparing benchmarks. Will clean jars..."
+    CLEAN="clean"
+    shift
+fi
 
 echo "Building RoaringBitmap jar"
-rm -f $BASEDIR/../target/RoaringBitmap*.jar
-mvn -f $BASEDIR/../pom.xml clean install -DskipTests -Dgpg.skip=true -Dcheckstyle.skip
+if [[ "$CLEAN" = "clean" ]]; then
+    rm -f $BASEDIR/../target/RoaringBitmap*.jar
+fi
+mvn -f $BASEDIR/../pom.xml $CLEAN install -DskipTests -Dgpg.skip=true -Dcheckstyle.skip
 
 [[ $? -eq 0 ]] || exit
 
 echo "Building Real Roaring Dataset jar"
-rm -f $BASEDIR/../real-roaring-dataset/target/real-roaring-dataset*.jar
-mvn -f $BASEDIR/../real-roaring-dataset/pom.xml clean install
+if [[ "$CLEAN" = "clean" ]]; then
+    rm -f $BASEDIR/../real-roaring-dataset/target/real-roaring-dataset*.jar
+fi
+mvn -f $BASEDIR/../real-roaring-dataset/pom.xml $CLEAN install
 
 [[ $? -eq 0 ]] || exit
 
 echo "Building benchmarks jar"
-rm -f  $BASEDIR/target/benchmarks.jar
-mvn -f $BASEDIR/pom.xml clean install -Dtest=*$1* -DfailIfNoTests=false -Dcheckstyle.skip
+if [[ "$CLEAN" = "clean" ]]; then
+    rm -f  $BASEDIR/target/benchmarks.jar
+fi
+mvn -f $BASEDIR/pom.xml $CLEAN install -Dtest=*${@:$#}* -DfailIfNoTests=false -Dcheckstyle.skip
 
 [[ $? -eq 0 ]] || exit
 
 echo "Running benchmarks"
-java -jar $BASEDIR/target/benchmarks.jar  true -wi 5 -i 5 -f 1 $@
+java -jar $BASEDIR/target/benchmarks.jar true -wi 5 -i 5 -f 1 $@
